+++ myems-web/src/components/MyEMS/Store/StoreEnergyCategory.js
@@ -1082,11 +1082,13 @@ const StoreEnergyCategory = ({ setRedirect, setRedirectUrl, t }) => {
         data={parameterLineChartData}
         options={parameterLineChartOptions}>
       </MultipleLineChart>
-      <br />
+
       <WorkingDaysConsumptionTable
        data={workingDaysConsumptionTableData}
        columns={workingDaysConsumptionTableColumns}>
       </WorkingDaysConsumptionTable>
+      <br />
+
       <DetailedDataTable data={detailedDataTableData} title={t('Detailed Data')} columns={detailedDataTableColumns} pagesize={50} >
       </DetailedDataTable>