+++ CHANGELOG.md
@@ -27,6 +27,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - updated translation of 'Description' in myems-admin
 - moved mapbox access token from CustomizeMapBox.js to config.js in myems-web, you can get access token at https://mapbox.com, if you wnat to turn off online map feature, please set showOnlineMap to false
 - updated WorkingDaysConsumptionTable in myems-web
+- updated FalconCardHeader in myems-web
+- 
 ### Fixed
 - fixed non-standard actions of microgrid in myems-api and myems-admin
 - fixed empty qrcode issue when creating objects in myems-admin